pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoscaling-analysis"
version = "0.1.0"
description = "Forecasting + autoscaling simulation pipeline (NASA logs) with Streamlit dashboard"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Your Team" }
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "pyarrow>=14.0",
  "scikit-learn>=1.3",
  "xgboost>=2.0",
  "matplotlib>=3.7",
  "streamlit>=1.32",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6",
  "black>=24.3",
  "mypy>=1.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["autoscaling_analysis*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = ["I"]  # import sorting
ignore = [
  "E501",  # allow long strings in config / regex
]
src = ["src", "scripts", "tests"]

[tool.ruff.lint.isort]
known-first-party = ["autoscaling_analysis"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
packages = ["autoscaling_analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"